Some final reference files.
Final Comment: This branch was a simple working branch for testing parallel development. New branches will now be created for more focused and refined refactorings, and this branch will be considered deprecated. The branch will not be deleted until a new parallel branch is finished, simply to allow any necessary benchmarking calculations.